requirements.txt
codegen==1.0
cron.py==0.0.5
Django==1.10.5
django-common-helpers==0.9.0
django-cron==0.5.0
django-datetime-widget==0.9.3
django-debug-toolbar==1.6
patterns==0.3
pytz==2016.10
sqlparse==0.2.2
wheel==0.24.0
